feat: prevent friend from being created if no friend request exists

+++ app/controllers/friends_controller.rb
@@ -2,10 +2,17 @@ class FriendsController < ApplicationController
   def create
     user = User.find_by(id: params[:user])
     friend = User.find_by(id: params[:friend])
+    friend_request = FriendRequest.find_by(friend_id: friend, user_id: user)
+
+    if friend_request.nil?
+      puts 'FRIEND REQUEST DOES NOT EXIST, YOU CANNOT BE FRIENDS SORRY'
+      redirect_back_or_to root_path
+      return
+    end
+
     puts 'THIS SUPPOSED TO BE ADDING A FREND'
 
     if user.friends << friend && friend.friends << user
-      friend_request = FriendRequest.find_sole_by(friend_id: friend, user_id: user)
       friend_request.destroy
 
       puts 'SAY HELLO, YOU ARE NOW FRENDS'
